pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ele"
version = "0.1.0"
description = "Pseudo-RAW preprocessing engine for Lightroom, Photoshop, and Capture One"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "tifffile>=2024.1",
    "typer>=0.12",
    "rich>=13.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    # Billing stack
    "sqlalchemy>=2.0",
    "stripe>=8.0",
    "bcrypt>=4.0",
    "python-dotenv>=1.0",
    "itsdangerous>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[project.scripts]
ele       = "ele.cli.main:app"
ele-admin = "ele.billing.admin:admin_app"

[tool.setuptools.packages.find]
where = ["."]
include = ["ele*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
